CompanyProduct

+++ Backend/StockWise/Dtos/RegisterDto.cs
@@ -5,6 +5,6 @@
         public string UserName {  get; set; }
         public string Email { get; set;}
         public string Password { get; set;}
-        public int CompanyNIP {  get; set;}
+        public long CompanyNIP {  get; set;}
     }
 }
